Depuracion y correccion de rutas, creacion de carpestas en view por rol user y admin
