+++ controllers/admin.Controller.js
@@ -1,9 +1,7 @@
 
 
 const admin = (req, res) => {
-    res.render('servc/admin',{
-        pagina:'Services'
-    })
+    res.render('admin/home')
 }
