+++ routes/admin.routes.js
@@ -1,9 +1,9 @@
 const express = require('express');
-const { admin } = require('../controllers/index.Controller.js');
+const { admin } = require('../controllers/admin.Controller.js');
 const isLoggedIn = require("../middleware/isLoggedIn");
 const router = express.Router();
 
 /* GET home page */
-router.get("/admin",isLoggedIn, admin);
+router.get("/",isLoggedIn, admin);
 
 module.exports = router;
