--- routes/home.routes.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-/* GET home page */
-router.get("/home", (req, res, next) => {
-  res.render("home");
-});
-
-module.exports = router;
